- NOTE: DownloadClient was renamed to FileDownloader.java in the previous two commits - Dagger was applied for FileDownloader.java
